move testAliasStarException to CalcValidationTest
